requirements.txt
python-telegram-bot>=20
openai>=1.40.0
gspread
google-auth
google-api-python-client
pydantic>=2.0.0
python-dotenv
tenacity
pytz
pillow
pypdf2
uvloop; platform_system=="Linux"
pytest
pytest-asyncio